analytics-service/requirements.txt
fastapi
uvicorn
clickhouse-connect
python-dateutil

